Fix 1 — sentinel type mismatch in FeatureMapping.cs:125: FeatureDistSquared now returns Double.MaxValue instead of Single.MaxValue for corner pairs of opposite convexity. Previously the float sentinel widened to ~3.4e38, which never matched the .Where(t => t.d != Double.MaxValue) filter in DoMapping, so convex↔concave pairs were never excluded from mapping candidates — they just sorted last. This matches upstream Kotlin's Float.MAX_VALUE return-and-filter behavior.
Fix 2 — corner reversal in Features.cs:141: CornerFeature's reversal now returns new CornerFeature(reversedCubics, !Convex) instead of an EdgeFeature, matching upstream Kotlin's Corner.reversed(). I also renamed the typo'd Reserved() to Reversed() across the Feature hierarchy (abstract declaration plus both overrides — nothing external called it, so the rename was safe).

New tests:

FeatureMappingTest.cs — FeatureDistSquaredIsMaxForOppositeConvexity asserts the sentinel is exactly Double.MaxValue in both argument orders; DoMappingExcludesOppositeConvexityPairs asserts a lone convex/concave candidate pair is filtered out so DoMapping falls back to the identity mapping (with the old bug this returned a (0.2, 0.7) pairing instead); FeatureMappingTriangleToStarHasNoOppositeConvexityPairs is an end-to-end check morphing a triangle against an 8-point star.
FeaturesTest.cs — ReversedCornerFlipsConvexityAndCubics verifies convex→concave and concave→convex with cubics reversed in both order and direction; ReversedEdgeStaysEdge covers the edge case.
Changes are uncommitted on the claude/practical-williams-c84de1 worktree branch — let me know if you'd like them committed.

+++ MaterialShapes/FeatureMapping.cs
@@ -122,7 +122,7 @@ public static class FeatureMapping
     internal static double FeatureDistSquared(Feature f1, Feature f2)
     {
         if (f1 is CornerFeature c1 && f2 is CornerFeature c2 && c1.Convex != c2.Convex)
-            return Single.MaxValue;
+            return Double.MaxValue;
 
         var p1 = FeatureRepresentativePoint(f1);
         var p2 = FeatureRepresentativePoint(f2);
